feat: 송금시 기록 저장하는 기능

+++ back-end/business/src/main/java/com/aicheck/business/domain/transaction_record/presentation/dto/TransactionRecordDto.java
@@ -2,25 +2,24 @@ package com.aicheck.business.domain.transaction_record.presentation.dto;
 
 import com.aicheck.business.domain.transaction_record.entity.TransactionRecord;
 import com.aicheck.business.domain.transaction_record.entity.TransactionType;
-
 import lombok.Builder;
 
 @Builder
 public record TransactionRecordDto(
-	String firstCategoryName,
-	String secondCategoryName,
-	TransactionType transactionType,
-	Integer amount,
-	Integer rating
+        String firstCategoryName,
+        String secondCategoryName,
+        TransactionType transactionType,
+        Long amount,
+        Integer rating
 ) {
 
-	public static TransactionRecordDto from(TransactionRecord transactionRecord) {
-		return TransactionRecordDto.builder()
-			.firstCategoryName(String.valueOf(transactionRecord.getFirstCategory()))
-			.secondCategoryName(String.valueOf(transactionRecord.getSecondCategory()))
-			.transactionType(transactionRecord.getType())
-			.amount(transactionRecord.getAmount())
-			.rating(transactionRecord.getRating())
-			.build();
-	}
+    public static TransactionRecordDto from(TransactionRecord transactionRecord) {
+        return TransactionRecordDto.builder()
+                .firstCategoryName(String.valueOf(transactionRecord.getFirstCategory()))
+                .secondCategoryName(String.valueOf(transactionRecord.getSecondCategory()))
+                .transactionType(transactionRecord.getType())
+                .amount(transactionRecord.getAmount())
+                .rating(transactionRecord.getRating())
+                .build();
+    }
 }
